cleanup relaycommand and move generic version to another file
